Calculo distancias estaciones
Código que calcula las distancias a las estaciones para cualquier rango dado. Se puede pulir un poco aun el programa (lo hare en estos días), y no estoy seguro de si los resultados que da están bien.
